new vision player table2

+++ Server/src/main/java/com/github/dr/rwserver/data/Player.java
@@ -44,7 +44,7 @@ public final class Player {
 	public long timeTemp;
 	public int ping = 50;
 	public int pingTimes=0;
-	public float avgPing=50;
+	public int avgPing=50;
 	public long lastMessageTime = 0;
 	public String lastSentMessage = "";
